refactor: make setitatr generic for document
