+++ docarray/base_document/document.py
@@ -9,14 +9,6 @@ from docarray.base_document.base_node import BaseNode
 from docarray.base_document.io.json import orjson_dumps, orjson_dumps_and_decode
 from docarray.base_document.mixins import IOMixin, UpdateMixin
 from docarray.typing import ID
-from docarray.utils._typing import is_tensor_union, is_type_tensor
-from docarray.utils.misc import is_tf_available
-
-tf_available = is_tf_available()
-if tf_available:
-    import tensorflow as tf  # type: ignore
-
-    from docarray.typing import TensorFlowTensor
 
 if TYPE_CHECKING:
     from docarray.array.stacked.array_stacked import DocumentArrayStacked
@@ -85,53 +77,23 @@ class BaseDocument(BaseModel, IOMixin, UpdateMixin, BaseNode):
 
     def __setattr__(self, name, value):
 
-        if self._da_ref is not None:
-            if name in self.__fields__:
-                field_type = self._get_field_type(name)
-                if tf_available and (
-                    issubclass(field_type, TensorFlowTensor)
-                    or isinstance(value, TensorFlowTensor)
-                    or isinstance(value, tf.Tensor)
-                ):
-                    raise RuntimeError(
-                        'Tensorflow tensor are immutable, Since this '
-                        'Document is stack you cannot change its '
-                        'tensor value. You should either unstack your '
-                        'DocumentArrayStacked or work at the '
-                        'DocumentArrayStack column level directly'
-                    )
+        if self._da_ref is not None and self._da_index is not None:
+            if name in self._da_ref.stacked_field():
+                old_value = getattr(self, name)
+                super().__setattr__(name, value)
+                # here cheat by calling setattr on the value but what we want to do is
+                # to call validation the same way pydantic does,
+                # but we actually don't want to set it
+                new_value = getattr(self, name)
+                try:
+                    self._da_ref.__update_columns__(name, self._da_index, value)
+                except Exception as e:
+                    object.__setattr__(self, name, old_value)
+                    raise e  # if something is not right when putting in
+                    # the da stacked we revert the change
 
-                elif is_type_tensor(field_type) or is_tensor_union(field_type):
-                    ### here we want to enforce the inplace behavior
-                    old_value = getattr(self, name)
-                    super().__setattr__(name, value)
-                    # here cheat by calling setattr on the value but what we want to do is
-                    # to call validation the same way pydantic does,
-                    # but we actually don't want to set it
-                    new_value = getattr(self, name)
-                    try:
-                        old_value[:] = new_value  # we force to do inplace
-                        new_value = old_value
-                    except Exception as e:
-                        object.__setattr__(self, name, old_value)
-                        raise e  # if something is not right when putting in
-                        # the da stacked we revert the change
+                object.__setattr__(self, name, new_value)
 
-                    object.__setattr__(self, name, new_value)
-
-                    return  # needed here to stop func execution
-
-                elif issubclass(field_type, BaseDocument):
-                    old_value = getattr(self, name)
-                    super().__setattr__(name, value)
-                    new_value = getattr(self, name)
-                    try:
-                        self._da_ref._doc_columns[name][self._da_index] = new_value
-                    except Exception as e:  # if something is not right when putting in
-                        # the da stacked we revert the change
-                        object.__setattr__(self, name, old_value)
-                        raise e
-
-                    object.__setattr__(self, name, new_value)
+                return  # needed here to stop func execution
 
         super().__setattr__(name, value)
